refactor: split sidebar into smaller components
